+++ src/com/jec/commons/jcad/spi/JcadValidator.ts
@@ -1,0 +1,47 @@
+//  DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
+//
+//   Copyright 2016-2017 Pascal ECHEMANN.
+//
+//   Licensed under the Apache License, Version 2.0 (the "License");
+//   you may not use this file except in compliance with the License.
+//   You may obtain a copy of the License at
+//
+//       http://www.apache.org/licenses/LICENSE-2.0
+//
+//   Unless required by applicable law or agreed to in writing, software
+//   distributed under the License is distributed on an "AS IS" BASIS,
+//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//   See the License for the specific language governing permissions and
+//   limitations under the License.
+
+import {JcadContextError} from "../exceptions/JcadContextError";
+
+/**
+ * A helper class that is used to validate JCAD configs.
+ */
+export class JcadValidator {
+
+  //////////////////////////////////////////////////////////////////////////////
+  // Constructor function
+  //////////////////////////////////////////////////////////////////////////////
+
+  /**
+   * Creates a new <code>JcadValidator</code> instance.
+   */
+  constructor() {}
+
+  //////////////////////////////////////////////////////////////////////////////
+  // Public methods
+  //////////////////////////////////////////////////////////////////////////////
+
+  /**
+   * Throws an error if the specified <code>context</code> parameter is 
+   * <code>null</code>.
+   * 
+   * @param {any} context the context to validate.
+   * @param {string} contextType the type of the context to validate.
+   */
+  public validateContext(context:any, contextType:string):void {
+    if(!context) throw new JcadContextError(`${contextType} must not be null!`);
+  }
+}